# contracts/openapi.json
{"openapi":"3.0.1","info":{"title":"Payment Gateway API","description":"Frontend-facing gateway for AI token wallet, pricing, demo purchase, and estimate APIs.","contact":{"name":"Jobseeker Copilot"},"license":{"name":"MIT"},"version":"2.0.0"},"servers":[{"url":"http://localhost","description":"Generated server url"}],"paths":{"/api/v1/payment/estimate":{"post":{"tags":["payment-gateway-controller"],"summary":"Estimate AI token usage","operationId":"estimate","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/EstimateRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/EstimateResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payment/demo-purchase":{"post":{"tags":["payment-gateway-controller"],"summary":"Demo purchase AI tokens","operationId":"demoPurchase","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DemoPurchaseRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DemoPurchaseResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payment/checkout":{"post":{"tags":["payment-gateway-controller"],"summary":"Create a Stripe checkout session for an AI token bundle","operationId":"checkout","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CheckoutRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CheckoutResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payment/wallet":{"get":{"tags":["payment-gateway-controller"],"summary":"Get the current user's AI token wallet","operationId":"wallet","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/WalletSummaryResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payment/transactions":{"get":{"tags":["payment-gateway-controller"],"summary":"Get the current user's AI token transactions","operationId":"transactions","parameters":[{"name":"X-Payment-Owner","in":"header","required":true},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":20}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/TransactionsResponse"}}}}},"security":[{"serviceToken":[]}]}},"/api/v1/payment/pricing":{"get":{"tags":["payment-gateway-controller"],"summary":"Get active AI token pricing plans","operationId":"pricing","parameters":[{"name":"X-Payment-Owner","in":"header","required":true}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PricingPlansResponse"}}}}},"security":[{"serviceToken":[]}]}}},"components":{"schemas":{"EstimateRequest":{"required":["feature"],"type":"object","properties":{"feature":{"type":"string"},"estimatedInputTokens":{"minimum":0,"type":"integer","format":"int64"},"estimatedOutputTokens":{"minimum":0,"type":"integer","format":"int64"}}},"EstimateResponse":{"type":"object","properties":{"feature":{"type":"string"},"estimatedTotalTokens":{"type":"integer","format":"int64"},"estimatedCostTokens":{"type":"integer","format":"int64"},"userBalanceTokens":{"type":"integer","format":"int64"},"canAfford":{"type":"boolean"}}},"DemoPurchaseRequest":{"required":["pricingPlanId"],"type":"object","properties":{"pricingPlanId":{"type":"string"}}},"DemoPurchaseResponse":{"type":"object","properties":{"wallet":{"$ref":"#/components/schemas/WalletSummaryResponse"},"transaction":{"$ref":"#/components/schemas/TransactionResponse"}}},"TransactionResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"transactionType":{"type":"string"},"tokenAmount":{"type":"integer","format":"int64"},"balanceBefore":{"type":"integer","format":"int64"},"balanceAfter":{"type":"integer","format":"int64"},"description":{"type":"string"},"referenceType":{"type":"string"},"referenceId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"}}},"WalletSummaryResponse":{"type":"object","properties":{"userId":{"type":"string"},"balanceTokens":{"type":"integer","format":"int64"},"lifetimePurchasedTokens":{"type":"integer","format":"int64"},"lifetimeSpentTokens":{"type":"integer","format":"int64"},"lifetimeRefundedTokens":{"type":"integer","format":"int64"},"freeTrialGranted":{"type":"boolean"}}},"CheckoutRequest":{"required":["pricingPlanId"],"type":"object","properties":{"pricingPlanId":{"type":"string"}}},"CheckoutResponse":{"type":"object","properties":{"sessionId":{"type":"string"},"checkoutUrl":{"type":"string"}}},"TransactionsResponse":{"type":"object","properties":{"userId":{"type":"string"},"transactions":{"type":"array","items":{"$ref":"#/components/schemas/TransactionResponse"}}}},"PricingPlansResponse":{"type":"object","properties":{"plans":{"type":"array","items":{"$ref":"#/components/schemas/TokenPricingPlanResponse"}}}},"TokenPricingPlanResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"tokenAmount":{"type":"integer","format":"int64"},"priceGbpPence":{"type":"integer","format":"int64"}}}},"securitySchemes":{"serviceToken":{"type":"apiKey","description":"Dedicated Job Seeker Copilot BFF service identity.","name":"X-Service-Token","in":"header"}}}}